+++ CHANGELOG.rst
@@ -1,8 +1,9 @@
 Change Log
 =============
 
-[4.17.9] - 2024-11-12
+[4.17.9] - 2024-11-15
 -----------------------
+- [FIXED] fix error in computing cost recovery
 - [FIXED] fix error in computing pDemandElecPeak in InputData
 - [CHANGED] keep the model in memory
 - [CHANGED] default values of minimum output results